Update to Mednafen 0.9.38.7 (October 1, 2015)

+++ src/hw_cpu/huc6280/huc6280.h
@@ -290,7 +290,7 @@ class HuC6280
 		break;
 
 	  case GSREG_TIMV:
-		value = timer_value;
+		value = (uint8)timer_value;
 		break;
 
 	  case GSREG_TIML:
